+++ Assets/Deadscripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a0dbb90f945d7d94f8dfcaa472bdad15
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
