+++ Assets/Deadscripts/OneWayPlatform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ca33a8bef45e74348823a8998d8e2d33
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
